seperate download component
