+++ frontend/src/components/opportunity/OpportunityDownload.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { useTranslations } from "next-intl";
+import { Button, Link } from "@trussworks/react-uswds";
+
+import { USWDSIcon } from "src/components/USWDSIcon";
+
+type Props = {
+  nofoPath: string;
+};
+
+const OpportunityDownload = ({ nofoPath }: Props) => {
+  const t = useTranslations("OpportunityListing.description");
+
+  const downloadNOFO = () => {
+    window.open(nofoPath, "_blank");
+  };
+
+  return (
+    <>
+      {nofoPath.length > 0 ? (
+        <div className="grid-row flex-justify">
+          <Button onClick={downloadNOFO} type="button">
+            <span>{t("nofo_download")} </span>
+            <USWDSIcon
+              name={"file_download"}
+              className="usa-icon usa-icon--size-4"
+            />
+          </Button>
+          <Link
+            className="flex-align-self-center"
+            href={"#opportunity_documents"}
+          >
+            {t("jump_to_documents")}
+          </Link>
+        </div>
+      ) : null}
+    </>
+  );
+};
+
+export default OpportunityDownload;
